Fix #745, out of bounds read in memcpy.
Entering `cmd /;=` and moving the cursor left twice resulted in an out
of bounds read.  Even moving the cursor left once triggered an assertion
in debug builds of Clink.

The word break enhancements from 7958647635 and d1555f3f6d malfunctioned
in an edge case where the cursor is in the last word but there are
non-word-break characters after the cursor.  The word break adjuster
logic had no way to know whether the line_state needed to stop at the
cursor vs at the end of the line, and it always assumed at the end of
the line.

Now line_state keeps track of whether the cursor or the end of the line
should limit the end of the last word.