Fix a long-standing issue in collect_words:
When the first word was an alias, then the tokeniser started on the next
space.  But if the rest of the line was spaces, then the loop would
incorrectly add an empty word, which would make argmatcher consume an
extra argument slot by mistake when parsing an expanded doskey alias.

The tokeniser needs to start at the next non-space character following
the command word.

There was even a unit test that had detected the mistake, but I didn't
know enough at the time to understand whether it was by design.