Add a mechanism to send a terminal request and read a response.
Terminal requests cause Windows Terminal to clear all queued input,
so this might not be something that can be used practically.

But this is able to send a terminal request sequence and read a
terminal report sequence matching a pattern.  If available input
doesn't match the pattern, then the processed input records are
re-queued and processed again as normal input.

The pattern is matched literally, except that '0' is a wildcard
meaning "one or more digits".