Fix match generation optimization with onadvance.
Clink caches generated matches for the end word.  There are rules for
when to discard the cached matches, and scripts can influence the rules.

But `onadvance` callbacks are another mechanism that can dynamically
change which argmatcher and arg slot generate matches for the line, and
there isn't enough context for scripts to overtly influence that (nor
do I think it should be scripts' responsibility).

This change automatically marks matches volatile (don't cache them) when
an `onadvance` callback is encountered in the end word.

An example of an argmatcher affected by this is `procdump`.  Its
-r and -e flags each accept an _optional_ argument, so depending on
what's typed a different match generator is used.  If the matches are
cached then invoking completion and typing something and invoking
completion again could use the cached matches instead of generating new
matches.