Improve Readline's typeahead optimization.
Readline disables its typeahead optimization while macro input is
pending.  But macro input is slow without the typeahead optimization.

The typeahead optimization seems to support macro input just fine *IF*
the host has an rl_input_available_hook callback that returns true when
macro input is available.

Furthermore, it looks like Readline doesn't recognize macro input in
general while checking for queued input.  It has a handful of special
cases or optimizations which are explicitly disabled when macro input is
present.  It looks like the logic works, and Readline just hasn't been
updated to enable the logic.

So, this change enables it for macro input.