Fix #751; multi-line prompt without newlines displays wrongly.
Readline had a bug in its "otherwise-blank last line" compensation logic
in _rl_update_final().  When replacing Readline's built-in display with
the new display_manager, I fixed that bug.  But the fix had a bug of its
own, and malfunctioned in the edge case where a prompt takes multiple
lines but contains no newline characters.

This change fixes that edge case.