Add `clink config theme save -d` flag.
Normally saving a *.clinktheme file saves the current values of color
settings, even if they match the default value.

The -d flag saves the current values of color settings, but if a value
for a color setting matches its default value then a note is written so
that loading the *.clinktheme file will reset to whatever the current
default value is at the moment when the file is loaded.