Fix #772; uninstall can fail to remove Clink from AutoRun.
If the logged on user is a non-admin and elevates, then it couldn't
remove Clink from AutoRun for the logged on user.  The installer has no
idea who the logged on user is; it only knows the elevated account.

There is a still a quirk where if the user is a non-admin and elevates,
then the installer cannot configure Clink for autorun for the current
user.  The reason is the same -- the installer only knows about the
elevated account, not who the logged on user is.