Fix #782; Win8.1 wrap behavior can add blank lines in some multiline prompts.
Windows 8.1 (and earlier) immediately wraps to the next line once a
character is printed at the end of a terminal line.  This results in an
extra blank line if the next character printed is a CR or LF.

Compensate by moving the cursor up a line first.

Another solution could have been to make ecma48_terminal_out handle the
case transparently.  But that would be a bigger and more complicated
change, especially since display_manager compensates for the Win8.1
autowrap issue in other places already.