Add troubleshooting code to the omp prompt script.
Oh-my-posh generates a Lua script that doesn't have error handling, so
when something goes wrong the real error message is swallowed and
instead a Lua syntax error is reported.

This change checks for a specific version of the generated init script
and replaces it with an updated version that includes error handling.

This should help trouble issue #790.