Fix halfway-active suggestion list after ALT-=.
Any command that prints output and shows a new prompt line left the
suggestion list in a partially-active state where it was not displayed,
but it responded to input, and UP/DOWN triggered regenerating new
suggestions.

There were two options:
1. Make it fully active and displayed.
2. Make it fully hidden and inactive until the input line changes.

I chose the second option because typically the printed output is for
immediate consumption by the user.  Printing the suggestion list again
interferes with implicit workflow, and can also scroll some of the
output out of view.