Fix pointer arithmetic mistake in unit tests.
line_editor_tester uses test_terminal_in to provide test input, but the
test_terminal_in::read method unconditionally increments the m_read
pointer.  Depending on the compiler and runtime state, this can
potentially lead to a buffer over-read and cause tests to fail.

It's been incorrect since it was first added in 2017, but it wasn't
always reachable in practice:

- Clink v1.1.1-alpha merged Readline 8.0, which added a typeahead
  optimization and made the test_terminal_in::read bug reachable.
- Clink v1.3.44 disabled Readline's typeahead optimization because of
  various bugs, making the bug unreachable.
- Clink v1.5.2 fixed the bugs and re-enabled Readline's typeahead
  optimization, making the bug reachable again.

Now it's fixed; test_terminal_in::read only advances the m_read pointer
if it hasn't reached the end of the test input.