Fix peeking macro input.
Discovered while investigating #802.  It got stuck in an infinite loop
thinking there was more input available, but the macro was finished but
rl_executing_macro hadn't been cleared yet, so the peek got confused.

The inaccurate peek was introduced more than 3 years ago, so it seems
unlikely to be related to issue #802, which was stated as starting in
the last few months.