Add CLINK_ANSI_HOST env var which overrides `terminal.emulation`.
When CLINK_ANSI_HOST is set, then it specifies the terminal support
available to Clink, and the `terminal.emulation` setting has no effect.

This is useful for forcing the terminal support when hosted inside an
embedded terminal, without needing to modify any profile settings (the
current terminal is a per-instance state, not a per-profile state).